requirements.txt
streamlit
python-dotenv
langchain-community
langchain-groq